prosper202: rename Prosper202\Bandit namespace + dir -> Lpo
