+++ 202-config/Lpo/DimensionSync.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Prosper202\Bandit;
+namespace Prosper202\Lpo;
 
 use mysqli;
 use Prosper202\Database\Connection;
