+++ 202-cronjobs/bandit_dimensions.php
@@ -36,8 +36,8 @@ error_reporting(E_ALL);
 
 include_once(str_repeat("../", 1) . '202-config/connect.php');
 
-use Prosper202\Bandit\DimensionSync;
-use Prosper202\Bandit\PairingClient;
+use Prosper202\Lpo\DimensionSync;
+use Prosper202\Lpo\PairingClient;
 use Prosper202\Database\Connection;
 
 set_time_limit(0);
