+++ tracking202/setup/aff_campaigns.php
@@ -234,7 +234,7 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 		// Landing Page Optimizer (segments-v2 G10): flag this user's dimension
 		// snapshot dirty; the hourly cron pushes it. DB-only — no HTTP here.
 		// (After the public-id block above: markDirty's queries reset insert_id.)
-		\Prosper202\Bandit\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
+		\Prosper202\Lpo\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
 
 		$_GET['copy_aff_campaign_id'] = false;
 	}
@@ -264,7 +264,7 @@ if (isset($_GET['delete_aff_campaign_id'])) {
 			// Landing Page Optimizer (segments-v2 G10): deletes change the
 			// synced snapshot too (buildSnapshot omits deleted rows) — flag
 			// dirty so the hourly cron re-pushes. DB-only — no HTTP here.
-			\Prosper202\Bandit\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
+			\Prosper202\Lpo\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
 		}
 	} else {
 		header('location: ' . get_absolute_url() . 'tracking202/setup/aff_campaigns.php');
