+++ tracking202/setup/aff_networks.php
@@ -80,7 +80,7 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 
 		// Landing Page Optimizer (segments-v2 G10): flag this user's dimension
 		// snapshot dirty; the hourly cron pushes it. DB-only — no HTTP here.
-		\Prosper202\Bandit\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
+		\Prosper202\Lpo\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
 
 		if ($slack) {
 			if ($editing == true) {
@@ -143,7 +143,7 @@ if (isset($_GET['delete_aff_network_id'])) {
 			// Landing Page Optimizer (segments-v2 G10): mirror the save-path
 			// hook above — every dictionary mutation on this page flags the
 			// snapshot dirty for the hourly cron. DB-only — no HTTP here.
-			\Prosper202\Bandit\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
+			\Prosper202\Lpo\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
 
 			if ($slack)
 				$slack->push('campaign_category_deleted', ['name' => $_GET['delete_aff_network_name'], 'user' => $user_row['username']]);
