+++ tracking202/setup/landing_pages.php
@@ -206,7 +206,7 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 		// Landing Page Optimizer (segments-v2 G10): flag this user's dimension
 		// snapshot dirty; the hourly cron pushes it. DB-only — no HTTP here.
 		// (After the public-id block above: markDirty's queries reset insert_id.)
-		\Prosper202\Bandit\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
+		\Prosper202\Lpo\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
 	}
 }
 
@@ -234,7 +234,7 @@ if (isset($_GET['delete_landing_page_id'])) {
 			// Landing Page Optimizer (segments-v2 G10): deletes change the
 			// synced snapshot too (buildSnapshot omits deleted rows) — flag
 			// dirty so the hourly cron re-pushes. DB-only — no HTTP here.
-			\Prosper202\Bandit\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
+			\Prosper202\Lpo\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
 
 			if ($slack) {
 				if (isset($_GET['delete_landing_page_type']) && $_GET['delete_landing_page_type'] == '0') {
@@ -499,7 +499,7 @@ template_top('Landing Page Setup');  ?>
 						} catch (Throwable $bandit_lookup_error) {
 							// pre-upgrade schema; keep the panel link
 						}
-						$bandit_create_base = \Prosper202\Bandit\PairingClient::saasBaseUrl() . '/api/customers/experiments/create';
+						$bandit_create_base = \Prosper202\Lpo\PairingClient::saasBaseUrl() . '/api/customers/experiments/create';
 
 						$landing_page_sql = "SELECT * FROM `202_landing_pages` WHERE `user_id`='" . $mysql['user_id'] . "' AND landing_page_type='1' AND landing_page_deleted='0'";
