+++ tracking202/setup/ppc_accounts.php
@@ -79,7 +79,7 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 			$add_success = true;
 			// Landing Page Optimizer (segments-v2 G10): flag this user's dimension
 			// snapshot dirty; the hourly cron pushes it. DB-only — no HTTP here.
-			\Prosper202\Bandit\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
+			\Prosper202\Lpo\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
 			if ($network_editing == true) {
 				if ($slack)
 					$slack->push('traffic_source_name_changed', ['old_name' => $_GET['edit_ppc_network_name'], 'new_name' => $_POST['ppc_network_name'], 'user' => $user_row['username']]);
@@ -177,7 +177,7 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 			// Landing Page Optimizer (segments-v2 G10): flag this user's dimension
 			// snapshot dirty; the hourly cron pushes it. DB-only — no HTTP here.
 			// (After the insert_id capture above: markDirty's queries reset it.)
-			\Prosper202\Bandit\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
+			\Prosper202\Lpo\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
 
 			foreach ($_POST['pixel_type_id'] as $key => $value) {
 				$mysql['pixel_type_id'] = $db->real_escape_string($value);
@@ -273,7 +273,7 @@ if (isset($_GET['delete_ppc_network_id'])) {
 			// Landing Page Optimizer (segments-v2 G10): deletes change the
 			// synced snapshot too (buildSnapshot omits deleted rows) — flag
 			// dirty so the hourly cron re-pushes. DB-only — no HTTP here.
-			\Prosper202\Bandit\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
+			\Prosper202\Lpo\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
 			if ($slack)
 				$slack->push('traffic_source_deleted', ['name' => $_GET['delete_ppc_network_name'], 'user' => $user_row['username']]);
 		}
@@ -305,7 +305,7 @@ if (isset($_GET['delete_ppc_account_id'])) {
 			// Landing Page Optimizer (segments-v2 G10): deletes change the
 			// synced snapshot too (buildSnapshot omits deleted rows) — flag
 			// dirty so the hourly cron re-pushes. DB-only — no HTTP here.
-			\Prosper202\Bandit\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
+			\Prosper202\Lpo\DimensionSync::markDirty($db, (int) ($_SESSION['user_id'] ?? 0));
 			if ($slack)
 				$slack->push('traffic_source_account_deleted', ['account_name' => $_GET['delete_ppc_account_name'], 'user' => $user_row['username']]);
 		}
